+++ NEWS.md
@@ -1,3 +1,9 @@
+# countries 1.1.0
+* Added function *list_countries* to obtain a list of countries
+* Added function *random_countries* to obtain random country names
+* Corrected small nomenclature mistake in the country reference list
+* Added two new palettes to *palettes_countries*
+* Updated vignettes
 
 # countries 1.0.5
 * Patch to address change in behaviour of function *is.atomic*